requirements.txt
robotframework>=4.1.2
jmespath>=1.0.1
python-dateutil>=2.8.2
requests>=2.31.0
pyyaml 
docker 
azure-containerregistry 
azure-identity 